backend/requirements.txt
# #Intel
# Flask==3.0.3
# Flask_Cors==4.0.0
# librosa==0.10.1
# matplotlib==3.8.3
# numpy==1.26.4
# pandas==2.2.2
# scikit_learn==1.4.1.post1
# tensorflow==2.15.0
# tensorflow_hub==0.16.0
# # tensorflow_intel==2.15.0

# MACOS SILICON
# Flask==3.0.3
# Flask_Cors==4.0.0
# librosa==0.10.1
# matplotlib==3.8.3
# numpy>=1.21,<2.0  
# pandas==2.2.2
# scikit_learn==1.4.1.post1
# tensorflow-macos==2.15.0
# tensorflow-metal==1.1.0
# tensorflow_hub==0.16.0
# # scipy>=1.7.0
# scipy>=1.10.0,<1.14.0
# reportlab==4.1.0
# shap==0.44.1
# tensorflow-probability==0.23.0
# transformers>=4.21.0
# transformers>=4.21.0 

# Core requirements
Flask==3.0.3
Flask_Cors==4.0.0
librosa==0.10.1
matplotlib==3.8.3
numpy>=1.21,<2.0  
pandas==2.2.2
scikit_learn==1.4.1.post1
# Cross-platform tensorflow
tensorflow==2.15.0
tensorflow_hub==0.16.0
scipy>=1.10.0,<1.14.0
reportlab==4.1.0
shap==0.44.1
tensorflow-probability==0.23.0
transformers>=4.21.0

# Deployment requirements
gunicorn==20.1.0
python-dotenv==1.0.1 